move QQ::Logging -> QQ::Config::Logging
Because QQ::Logger would be confusingly adjacent, and this class is really
a configuration object
